+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to the FolioBlocks project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2026-08-05
+### Fixed
+- Fixed bug in Filmstrip Gallery that caused a fatal error on free version of 1.5.0
+- Fixed bug in Video Block that caused undefined variable error on free version of 1.5.0
+
 ## [1.5.0] - 2026-08-04
 ### Changed
 - Refinements to the Proofing Session page.